gpu: move array-intrinsic inlining into its own unit
all, matmul, sum, dot_product and transpose all have to become explicit
loops before a kernel can hold them, and the machinery they share -- the
section bounds a mask is walked for, the elementwise rewrite of a mask,
the search for an array intrinsic inside a larger expression -- is used
by nothing else. Together they were more than a fifth of the file.

Pure code motion: the definitions are unchanged, only their home is.
